Transform public/module/k8s/index.ts into re-export-only module
